Add files via upload
This file is the original analytics process model for CRISP-DM. This process model is one of the most highly adopted process approaches for Data Science. It was originally available on the internet from this location - https://www.the-modeling-agency.com/crisp-dm.pdf. In 2022, the website lapsed leaving links to this location disabled. The original site location and document is available from the Internet Wayback Machine - https://web.archive.org/web/20220401041957/https:/www.the-modeling-agency.com/crisp-dm.pdf. I am also hosting the file in this GitHub location to retain a permanent link to the PDF document.